Removed the LinuxAmd64 transpiler and backend; made LLVM the new one.
