--- src/assembler/nasm/nasmAssemblerTarget.ts
@@ -1,4 +1,0 @@
-export enum NasmAssemblerTarget
-{
-    Elf64 = 'elf64',
-}
