+++ src/main.ts
@@ -9,7 +9,6 @@ import FileSystem from 'fs';
 import { Importer } from './importer/importer';
 import { Lexer } from './lexer/lexer';
 import { LinuxAmd64Backend } from './backends/linuxAmd64Backend';
-import { LinuxAmd64LlvmBackend } from './backends/linuxAmd64LlvmBackend';
 import { Lowerer } from './lowerer/lowerer';
 import os from 'os';
 import { Parser } from './parser/parser';
@@ -37,7 +36,7 @@ class Main
 
     public run (): void
     {
-        const standardLibraryTargetPath = this.getStandardLibraryPath();
+        const standardLibraryTargetPath = Path.join(this.arguments.standardLibraryPath, this.arguments.targetPlatform);
 
         const diagnostic = new Diagnostic.Diagnostic();
 
@@ -124,26 +123,7 @@ class Main
                 backend.run(intermediateLanguage, standardLibraryFilePath, temporaryDirectoryPath, this.arguments.outputPath);
                 break;
             }
-            case TargetPlatform.LinuxAmd64Llvm:
-            {
-                const backend = new LinuxAmd64LlvmBackend();
-                backend.run(intermediateLanguage, standardLibraryFilePath, temporaryDirectoryPath, this.arguments.outputPath);
-                break;
-            }
         }
-    }
-
-    private getStandardLibraryPath (): string
-    {
-        let platformPath = this.arguments.targetPlatform;
-
-        if (platformPath == TargetPlatform.LinuxAmd64Llvm)
-        {
-            platformPath = TargetPlatform.LinuxAmd64;
-        }
-
-        const standardLibraryTargetPath = Path.join(this.arguments.standardLibraryPath, platformPath);
-        return standardLibraryTargetPath;
     }
 }
