+++ src/options/targetPlatform.ts
@@ -1,5 +1,4 @@
 export enum TargetPlatform
 {
     LinuxAmd64 = 'linuxAmd64',
-    LinuxAmd64Llvm = 'linuxAmd64Llvm',
 }
