--- src/transpiler/amd64/linux/directiveInstruction.ts
@@ -1,19 +1,0 @@
-import * as Instructions from '../../common/instructions';
-
-/**
- * An instruction with its render options set according to NASM Assembler directives
- */
-export class DirectiveInstruction extends Instructions.Instruction
-{
-    constructor (command: string, ...operands: string[])
-    {
-        super(command, ...operands);
-
-        this.renderOptions = {
-            commandOperandSplitter: ' ',
-            operandSplitter: ' ',
-            prefix: '[',
-            postfix: ']',
-        };
-    }
-}
