--- src/transpiler/amd64/registers/register8Amd64.ts
@@ -1,9 +1,0 @@
-export class Register8Amd64
-{
-    public readonly bit8: string;
-
-    constructor (bit8: string)
-    {
-        this.bit8 = bit8;
-    }
-}
